attributes

+++ src/main/java/entities/Attributes.java
@@ -13,16 +13,16 @@ public class Attributes implements Serializable {
     private List<String> breed; // allow for multiple preferred breeds
     private List<Integer> age; // allow for a range of preferred ages
     private String gender;
-    private int preferredProximity; // distance in KM
+    private float preferredProximity; // distance in KM
     private Image proofOfVaccination;
     private boolean vaccineStatus;
-    private int longitude;
-    private int latitude;
+    private float longitude;
+    private float latitude;
 
     private List<DayOfWeek> availableDay; // list of the available days of the week
 
     public Attributes(List<String> species, List<String> breed, List<Integer> age, String gender, Image proofOfVaccination,
-                      boolean vaccineStatus, int longitude, int latitude, int preferredProximity, List<DayOfWeek> availableDay){
+                      boolean vaccineStatus, float longitude, float latitude, float preferredProximity, List<DayOfWeek> availableDay){
         this.species = species;
         this.breed = breed;
         this.age = age;
@@ -58,14 +58,14 @@ public class Attributes implements Serializable {
         this.vaccineStatus = vaccine_status;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
-    public void setPreferredProximity(int preferredProximity) {
+    public void setPreferredProximity(float preferredProximity) {
         this.preferredProximity = preferredProximity;
     }
 
@@ -96,15 +96,15 @@ public class Attributes implements Serializable {
         return vaccineStatus;
     }
 
-    public int getPreferredProximity() {
+    public float getPreferredProximity() {
         return preferredProximity;
     }
 
-    public int getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public int getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
